% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{draw_circles}
\alias{draw_circles}
\title{Plot circles}
\usage{
draw_circles(x, y, r, polygon_args = list(), vertices = 500, add = FALSE,
  asp = 1, ...)
}
\arguments{
\item{x}{x Coordinates for the circles' centers}

\item{y}{y coordinates for the circles' centers}

\item{r}{Radiuses of the circles}

\item{polygon_args}{List of arguments to pass to
\code{\link[graphics]{polygon}} to adjust borders, fill, shadings, etc.}

\item{vertices}{The number of vertices used for drawing the circumference of
each circle.}

\item{add}{If \code{FALSE} initializes a new plotting frame with
\code{\link[graphics]{plot}}.}

\item{asp}{Controls the aspect ratio of the plot. Will not produce circles
if it ses to anything other than 1.}

\item{\dots}{Arguments to pass to \code{\link[graphics]{plot}}}
}
\value{
Plots circles either in the current plotting window or on a new
    plotting window.
}
\description{
Plots any numbers of circles using base R graphics, starting by plotting the
first circle on the bottom and going up.
}
\details{
draw_circles uses \code{\link[graphics]{polygon}} to draw the circles,
which means that they are technically just circle-shaped polygons. The
function is vectorized and can handle any number of circles.
}
\examples{
x <- runif(3)
y <- runif(3)
r <- runif(3, 0.2, 0.6)

draw_circles(x, y, r, polygon_args = list(border = "transparent",
                                          col = c("goldenrod1",
                                                  "steelblue1",
                                                  "black"))
)
}
\seealso{
\code{\link[graphics]{polygon}}, \code{\link[graphics]{plot}}
}

