% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circles.R
\name{intersect_circles}
\alias{intersect_circles}
\title{Find intersection points of circles}
\usage{
intersect_circles(x, y, r, nan.rm = FALSE)
}
\arguments{
\item{x}{x coordinates of the circle centers}

\item{y}{y coordinates of the circle centers}

\item{r}{Radiuses of the circles}

\item{nan.rm}{Whether to drop}
}
\value{
Return a matrix of x and y coordinates as well as the indices of the
    circles.
}
\description{
Intersects any number of circles to find the points at which they intersect.
}
\examples{
r <- runif(5, 0.3, 0.8)
x <- runif(5)
y <- runif(5)

int <- intersect_circles(x, y, r)
int

draw_circles(x, y, r)
points(int[, 1], int[, 2], pch = 16)

}

