% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circles.R
\name{locate_intersections}
\alias{locate_intersections}
\title{Find intersection points of circles}
\usage{
locate_intersections(r1, r2, x_d, y_d, x_c, y_c, d)
}
\description{
Find intersection points of circles
}

