% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{farthest_points}
\alias{farthest_points}
\title{Find the farthest points}
\usage{
farthest_points(data, n)
}
\arguments{
\item{data}{A matrix of points to choose n from}

\item{n}{The number of points to select}
}
\value{
A vector of indices for the points in the subset
}
\description{
Select a subset of points that maximize the smallest pairwise distance
between the points
}
\details{
This function begins with a random sample of n points and iterately puts them
back into the mother set and tries to see if there is a better candidate than
the point that was put back.
}
\examples{
xy <- matrix(runif(200), ncol = 2)
id <- farthest_points(data = xy, n = 5)
xy[id, ]

plot(xy)
points(xy[id, ], pch = 16)

}
\seealso{
\code{\link[stats]{dist}}
}

